backend/requirements.txt
blinker==1.6.2
certifi==2023.5.7
click==8.1.4
dnspython==2.3.0
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pymongo==4.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
Werkzeug==2.3.6
